third_party/WebKit/LayoutTests/fast/dom/document-all-tags.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Test that HTMLAllCollection.tags() matches IE behavior with too few arguments.');
shouldThrow("document.all.tags()");
</script>
</body>
</html>
